--- frontend/src/classes/ValidationError.js
@@ -1,6 +1,0 @@
-export class ValidationError extends Error {
-  constructor(message, type) {
-    super(message);
-    this.type = type;
-  }
-}
