+++ frontend/src/components/Common/Icons/ClearIcon.js
