+++ frontend/src/components/Common/LatestBlock.tsx
@@ -1,4 +1,5 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
+
 import { calculateBlockTimeDifference } from '@/utils/calculateBlockTimeDifference';
 
 interface LatestBlockProps {
