+++ frontend/src/components/Editor/EditorComponents/GlyphContainer.js
