+++ frontend/src/components/Editor/EditorComponents/MonacoEditorComponent.jsx
