+++ frontend/src/components/Editor/EditorComponents/custom.d.ts
@@ -1,0 +1,4 @@
+declare module '!!raw-loader!*' {
+  const content: string;
+  export default content;
+}
