+++ frontend/src/components/Editor/EditorViewContainer/DeveloperToolsContainer.tsx
@@ -1,19 +1,17 @@
 import React, { useContext } from 'react';
-import DeveloperToolsView from '../EditorView/DeveloperToolsView';
+
 import { IndexerDetailsContext } from '../../../contexts/IndexerDetailsContext';
+import DeveloperToolsView from '../EditorView/DeveloperToolsView';
 
 interface DeveloperToolsContainerProps {
   handleFormating: () => void;
   handleCodeGen: () => void;
-  executeIndexerFunction: () => void;
   isExecuting: boolean;
+  executeIndexerFunction: () => void;
+  heights: number[];
+  setHeights: React.Dispatch<React.SetStateAction<number[]>>;
   stopExecution: () => void;
-  heights: string[];
-  setHeights: React.Dispatch<React.SetStateAction<string[]>>;
-  latestHeight: number;
-  isUserIndexer: boolean;
-  handleDeleteIndexer: () => void;
-  error: string;
+  latestHeight: number | undefined;
   diffView: boolean;
   setDiffView: React.Dispatch<React.SetStateAction<boolean>>;
 }
@@ -27,24 +25,10 @@ const DeveloperToolsContainer: React.FC<DeveloperToolsContainerProps> = ({
   heights,
   setHeights,
   latestHeight,
-  isUserIndexer,
-  handleDeleteIndexer,
-  error,
   diffView,
   setDiffView,
 }) => {
-  const {
-    indexerName,
-    accountId,
-    indexerDetails,
-    setShowPublishModal,
-    setShowResetCodeModel,
-    setShowForkIndexerModal,
-    debugMode,
-    setDebugMode,
-    isCreateNewIndexer,
-    setShowLogsView,
-  } = useContext(IndexerDetailsContext);
+  const { setShowResetCodeModel, debugMode, setDebugMode } = useContext(IndexerDetailsContext);
 
   const removeHeight = (index: number): void => {
     setHeights(heights.filter((_, i) => i !== index));
@@ -53,20 +37,23 @@ const DeveloperToolsContainer: React.FC<DeveloperToolsContainerProps> = ({
   return (
     <DeveloperToolsView
       {...{
+        // Props
         handleFormating,
         handleCodeGen,
-        setShowResetCodeModel,
-        debugMode,
-        setDebugMode,
+        executeIndexerFunction,
+        isExecuting,
+        stopExecution,
         heights,
         setHeights,
         latestHeight,
-        isExecuting,
-        stopExecution,
-        removeHeight,
-        executeIndexerFunction,
         diffView,
         setDiffView,
+        // Context
+        setShowResetCodeModel,
+        debugMode,
+        setDebugMode,
+        // Functions
+        removeHeight,
       }}
     />
   );
