--- frontend/src/components/Editor/EditorViewContainer/EditorMenuContainer.jsx
@@ -1,50 +1,0 @@
-import React, { useContext } from 'react';
-import EditorMenuView from '../EditorView/EditorMenuView';
-import { IndexerDetailsContext } from '../../../contexts/IndexerDetailsContext';
-
-const EditorMenuContainer = (props) => {
-  const {
-    handleFormating,
-    handleCodeGen,
-    error,
-    executeIndexerFunction,
-    heights,
-    setHeights,
-    isCreateNewIndexer,
-    isExecuting,
-    stopExecution,
-    latestHeight,
-    isUserIndexer,
-    handleDeleteIndexer,
-  } = props;
-  const {
-    indexerName,
-    accountId,
-    indexerDetails,
-    setShowPublishModal,
-    setShowResetCodeModel,
-    setShowForkIndexerModal,
-    debugMode,
-    setShowLogsView,
-  } = useContext(IndexerDetailsContext);
-
-  return (
-    <EditorMenuView
-      {...props}
-      {...{
-        indexerName,
-        accountId,
-        indexerDetails,
-        setShowPublishModal,
-        setShowResetCodeModel,
-        setShowForkIndexerModal,
-        debugMode,
-        isCreateNewIndexer,
-        setShowLogsView,
-        handleDeleteIndexer,
-      }}
-    />
-  );
-};
-
-export default EditorMenuContainer;
