+++ frontend/src/components/Logs/LogsViewContainer/ClearButtonContainer.tsx
@@ -1,4 +1,5 @@
-import React, { type MouseEvent, type FC } from 'react';
+import React, { type FC, type MouseEvent } from 'react';
+
 import ClearButtonView from '../LogsView/ClearButtonView';
 
 interface ClearButtonProps {
