+++ frontend/src/components/Logs/LogsViewContainer/DateSelectorContainer.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
-import DateSelectorView from '../LogsView/DateSelectorView';
+
 import { TIME_INTERVALS_MAP } from '@/constants/DurationMap';
+
+import DateSelectorView from '../LogsView/DateSelectorView';
 
 interface DateSelectorProps {
   selectedDate: string;
