+++ frontend/src/components/Logs/LogsViewContainer/LogTypeSelectorContainer.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import LogTypeSelectorView from '../LogsView/LogTypeSelectorView';
 
 interface LogTypeSelectorContainerProps {
