+++ frontend/src/contexts/ModalContext.js
@@ -1,4 +1,4 @@
-import React, { createContext, useState, useContext } from 'react';
+import React, { createContext, useContext, useState } from 'react';
 
 const ModalContext = createContext({
   openModal: false,
