+++ frontend/src/pages/query-api-editor/index.js
@@ -1,7 +1,8 @@
-import React, { useEffect, useContext } from 'react';
 import { withRouter } from 'next/router';
+import React, { useContext, useEffect } from 'react';
 import { Alert } from 'react-bootstrap';
-import Editor from '@/components/Editor';
+
+import Editor from '@/components/Editor/EditorComponents/Editor';
 import IndexerLogsContainer from '@/components/Logs/LogsViewContainer/IndexerLogsContainer';
 import { IndexerDetailsContext } from '@/contexts/IndexerDetailsContext';
