+++ frontend/src/utils/formatters.test.js
@@ -1,4 +1,4 @@
-import { formatSQL, formatIndexingCode } from './formatters';
+import { formatIndexingCode, formatSQL } from './formatters';
 
 const inputSQL1 = `CREATE TABLE\n  "indexer_storage" (\n    "function_name" TEXT NOT NULL,\n    "key_name" TEXT NOT NULL,\n    "value" TEXT NOT NULL,\n    PRIMARY KEY ("function_name", "key_name")\n  )\n`;
 const expectedOutput1 = `CREATE TABLE
